Adapted examples to new COutDevice, updated documentation of CAN driver
git-svn-id: https://svn-eos.cs.uni-magdeburg.de/repos/Projects/AVR/halib/trunk@445 ac1e7a95-1e17-0410-ad1e-f2b06e490645

+++ examples/applications/ucframe.cpp
@@ -93,7 +93,7 @@ int main()
 		len = 200;
 		if(cframe.receive(framedaten, len))
 		{
-			cdev << "\n\rreceive " << len <<"\t";
+			cdev << "\n\rreceive " << (int32_t)len <<"\t";
 			for(uint8_t i = 0; i < len; i++ )
 				cdev.put(framedaten[i]);
 		}
