+++ examples/applications/ucframe2.cpp
@@ -75,7 +75,7 @@ void onframercv()
 {
     len = 200;
     cframe.receive(framedaten, len);
-    cdev << "\n\rreceive " << len <<"\t";
+    cdev << "\n\rreceive " << (int32_t)len <<"\t";
     for(uint8_t i = 0; i < len; i++ )
 	cdev.put(framedaten[i]);
 }
